Clean up. Fixing error type at context children
